Fix up query object to clone itself. Extract utility methods to Montage::Support class
