+++ lib/montage.rb
@@ -1,7 +1,6 @@
 require 'montage/version'
 require 'montage/client'
 require 'montage/query'
-require 'montage/relation'
 
 module Montage
 end
