+++ test/montage/query_test.rb
@@ -2,64 +2,6 @@ require File.dirname(__FILE__) + '/../minitest_helper.rb'
 require 'montage/query'
 
 class Montage::QueryTest < Minitest::Test
-  context "#nillify" do
-    setup do
-      @query = Montage::Query.new
-    end
-
-    should "return the string if not empty" do
-      assert_equal "foo", @query.nillify("foo")
-    end
-
-    should "return nil if the string is empty" do
-      assert_nil @query.nillify("")
-    end
-
-    should "return the number if it is not zero" do
-      assert_equal 9, @query.nillify(9)
-    end
-
-    should "return nil if the number is zero" do
-      assert_nil @query.nillify(0)
-    end
-  end
-
-  context "#is_i?" do
-    setup do
-      @query = Montage::Query.new
-    end
-
-    should "return false if a float is passed in" do
-      assert !@query.is_i?("1.2")
-    end
-
-    should "return false if a string is passed in" do
-      assert !@query.is_i?("foo")
-    end
-
-    should "return true if an integer is passed in" do
-      assert @query.is_i?("1")
-    end
-  end
-
-  context "#is_f?" do
-    setup do
-      @query = Montage::Query.new
-    end
-
-    should "return false if an integer is passed in" do
-      assert !@query.is_f?("1")
-    end
-
-    should "return false if a string is passed in" do
-      assert !@query.is_f?("foo")
-    end
-
-    should "return true if a float is passed in" do
-      assert @query.is_f?("1.2")
-    end
-  end
-
   context "#limit" do
     setup do
       @query = Montage::Query.new
